+++ samples/system-test/spanner.test.js
@@ -43,6 +43,12 @@ const dmlCmd = 'node dml.js';
 const datatypesCmd = 'node datatypes.js';
 const backupsCmd = 'node backups.js';
 const instanceCmd = 'node instance.js';
+const createTableWithForeignKeyDeleteCascadeCommand =
+  'node table-create-with-foreign-key-delete-cascade.js';
+const alterTableWithForeignKeyDeleteCascadeCommand =
+  'node table-alter-with-foreign-key-delete-cascade.js';
+const dropForeignKeyConstraintDeleteCascaseCommand =
+  'node table-drop-foreign-key-constraint-delete-cascade.js';
 
 const CURRENT_TIME = Math.round(Date.now() / 1000).toString();
 const PROJECT_ID = process.env.GCLOUD_PROJECT;
@@ -312,6 +318,25 @@ describe('Spanner', () => {
       output,
       new RegExp(`Created database ${DATABASE_ID} on instance ${INSTANCE_ID}.`)
     );
+  });
+
+  // update_database
+  it('should set database metadata', async () => {
+    const output = execSync(
+      `node database-update.js ${INSTANCE_ID} ${DATABASE_ID} ${PROJECT_ID}`
+    );
+    assert.match(
+      output,
+      new RegExp(
+        `Waiting for update operation for ${DATABASE_ID} to complete...`
+      )
+    );
+    assert.match(output, new RegExp(`Updated database ${DATABASE_ID}.`));
+    // cleanup
+    const [operation] = await instance
+      .database(DATABASE_ID)
+      .setMetadata({enableDropProtection: false});
+    await operation.promise();
   });
 
   describe('encrypted database', () => {
@@ -1346,6 +1371,52 @@ describe('Spanner', () => {
     );
     assert.include(output, 'Version retention period: 1d');
     assert.include(output, 'Earliest version time:');
+  });
+
+  it('should create a table with foreign key delete cascade', async () => {
+    const output = execSync(
+      `${createTableWithForeignKeyDeleteCascadeCommand} "${INSTANCE_ID}" "${DATABASE_ID}" ${PROJECT_ID}`
+    );
+    assert.match(
+      output,
+      new RegExp(`Waiting for operation on ${DATABASE_ID} to complete...`)
+    );
+    assert.match(
+      output,
+      new RegExp(
+        'Created Customers and ShoppingCarts table with FKShoppingCartsCustomerId'
+      )
+    );
+  });
+
+  it('should alter a table with foreign key delete cascade', async () => {
+    const output = execSync(
+      `${alterTableWithForeignKeyDeleteCascadeCommand} "${INSTANCE_ID}" "${DATABASE_ID}" ${PROJECT_ID}`
+    );
+    assert.match(
+      output,
+      new RegExp(`Waiting for operation on ${DATABASE_ID} to complete...`)
+    );
+    assert.match(
+      output,
+      new RegExp('Altered ShoppingCarts table with FKShoppingCartsCustomerName')
+    );
+  });
+
+  it('should drop a foreign key constraint delete cascade', async () => {
+    const output = execSync(
+      `${dropForeignKeyConstraintDeleteCascaseCommand} "${INSTANCE_ID}" "${DATABASE_ID}" ${PROJECT_ID}`
+    );
+    assert.match(
+      output,
+      new RegExp(`Waiting for operation on ${DATABASE_ID} to complete...`)
+    );
+    assert.match(
+      output,
+      new RegExp(
+        'Altered ShoppingCarts table to drop FKShoppingCartsCustomerName'
+      )
+    );
   });
 
   describe('leader options', () => {