+++ test/batch-transaction.ts
@@ -149,12 +149,14 @@ describe('BatchTransaction', () => {
       gaxOptions: GAX_OPTS,
       params: {},
       types: {},
+      dataBoostEnabled: true,
     };
 
     it('should make the correct request', () => {
       const fakeParams = {
         params: {a: 'b'},
         paramTypes: {a: 'string'},
+        dataBoostEnabled: true,
       };
 
       const expectedQuery = Object.assign({sql: QUERY.sql}, fakeParams);
@@ -304,6 +306,7 @@ describe('BatchTransaction', () => {
       keys: ['a', 'b'],
       ranges: [{}, {}],
       gaxOptions: GAX_OPTS,
+      dataBoostEnabled: true,
     };
 
     it('should make the correct request', () => {
@@ -311,6 +314,7 @@ describe('BatchTransaction', () => {
       const expectedQuery = {
         table: QUERY.table,
         keySet: fakeKeySet,
+        dataBoostEnabled: true,
       };
 
       const stub = sandbox.stub(batchTransaction, 'createPartitions_');
@@ -347,6 +351,30 @@ describe('BatchTransaction', () => {
 
     it('should make query requests for non-read partitions', () => {
       const partition = {sql: 'SELECT * FROM Singers'};
+      const stub = sandbox.stub(batchTransaction, 'run');
+
+      batchTransaction.execute(partition, assert.ifError);
+
+      const query = stub.lastCall.args[0];
+      assert.strictEqual(query, partition);
+    });
+
+    it('should make read requests for read partitions with data boost enabled', () => {
+      const partition = {table: 'abc', dataBoostEnabled: true};
+      const stub = sandbox.stub(batchTransaction, 'read');
+
+      batchTransaction.execute(partition, assert.ifError);
+
+      const [table, options] = stub.lastCall.args;
+      assert.strictEqual(table, partition.table);
+      assert.strictEqual(options, partition);
+    });
+
+    it('should make query requests for non-read partitions with data boost enabled', () => {
+      const partition = {
+        sql: 'SELECT * FROM Singers',
+        dataBoostEnabled: true,
+      };
       const stub = sandbox.stub(batchTransaction, 'run');
 
       batchTransaction.execute(partition, assert.ifError);